Custom Hook for No Internet
