+++ src/components/NoInternet/CustomNoInternet.jsx
@@ -1,0 +1,23 @@
+import { useEffect, useState } from 'react';
+
+const CustomNoInternet = () => {
+  const [isOnline, setIsOnline] = useState(navigator.onLine);
+
+  useEffect(() => {
+    const handleOnlineStatus = () => {
+      setIsOnline(navigator.onLine);
+    };
+
+    window.addEventListener('online', handleOnlineStatus);
+    window.addEventListener('offline', handleOnlineStatus);
+
+    return () => {
+      window.removeEventListener('online', handleOnlineStatus);
+      window.removeEventListener('offline', handleOnlineStatus);
+    };
+  }, []);
+
+  return (isOnline)
+};
+
+export default CustomNoInternet;
