+++ src/components/NoInternet/NoInternet.stories.js
@@ -1,0 +1,13 @@
+import  NoInternetComponent  from '../NoInternet';
+
+export default {
+  title: 'Edustipend/Components/NoInternet',
+  component: NoInternetComponent,
+  tags: ['autodocs']
+};
+
+export const DefaultText = {
+  args: {
+    content: 'No Internet'
+  }
+};
