+++ src/components/SideBar/SideBar.stories.js
@@ -1,0 +1,13 @@
+import SideBarComponent from "../SideBar"
+
+export default {
+  title: 'Edustipend/Components/SideBar',
+  component: SideBarComponent,
+  tags: ['autodocs']
+};
+
+export const DefaultText = {
+  args: {
+    content: "Learner's Dashboard Side Bar"
+  }
+};
